backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Database (SQLite - no additional drivers needed)
sqlalchemy==2.0.36
alembic==1.14.0

# Utilities
python-dotenv==1.0.1
slowapi==0.1.9
